removed string
